+++ sdk/loganalytics/azure-mgmt-loganalytics/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 0.0.0 (2022-04-08)
+
+
+
 ## 13.0.0b3 (2022-04-07)
 
 **Features**
