+++ sdk/loganalytics/azure-mgmt-loganalytics/azure/mgmt/loganalytics/_configuration.py
@@ -19,8 +19,8 @@ if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential
 
 
-class OperationalInsightsManagementClientConfiguration(Configuration):
-    """Configuration for OperationalInsightsManagementClient.
+class LogAnalyticsManagementClientConfiguration(Configuration):
+    """Configuration for LogAnalyticsManagementClient.
 
     Note that all parameters used to create this instance are saved as instance
     attributes.
@@ -37,7 +37,7 @@ class OperationalInsightsManagementClientConfiguration(Configuration):
         subscription_id: str,
         **kwargs: Any
     ) -> None:
-        super(OperationalInsightsManagementClientConfiguration, self).__init__(**kwargs)
+        super(LogAnalyticsManagementClientConfiguration, self).__init__(**kwargs)
         if credential is None:
             raise ValueError("Parameter 'credential' must not be None.")
         if subscription_id is None:
