+++ sdk/loganalytics/azure-mgmt-loganalytics/azure/mgmt/loganalytics/_version.py
@@ -6,4 +6,4 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-VERSION = "13.0.0b3"
+VERSION = "0.0.0"
